refactor: utils 함수 관심사 별 폴더 구조 분리
