+++ src/utils/coupon/couponSorter.ts
@@ -1,4 +1,4 @@
-import { Coupon } from '../types/Coupon.type';
+import { Coupon } from '../../types/Coupon.type';
 
 export const sortCouponsByDiscountRate = (coupons: Coupon[]) => {
   return coupons
